refactor(FlatDB): widen arena/writer side to long sizes
Widens the write-path size types from int to long so that a single
ArenaReservation / SnapshotLocation can describe a >2 GiB region
once the WholeReadSession side of the read path becomes chunk-aware.
Concretely:

- IByteBufferWriter.Written: int → long (cascades through StreamBuffer,
  SpanBuffer, PooledByteBuffer writers).
- HSST/B-search builders now hold _baseOffset / _writtenBeforeValue /
  _startWritten as long; per-HSST deltas (≤2 GiB) are explicit (int)
  casts at the use sites.
- IArenaManager.CreateWriter / CompleteWrite / Touch take long sizes;
  ArenaWriter.Complete reports long; ArenaReservation.Size,
  PersistedSnapshot.Size, SnapshotLocation.Size are long;
  ArenaFile.{GetSpan,OpenWholeView,Touch,AdviseDontNeed} take long size.
- MmapWholeView holds size as long. GetSpan() still returns
  ReadOnlySpan<byte> and therefore casts checked((int)size) — for >2 GiB
  reservations callers must use a chunk-aware reader (a future evolution
  of WholeReadSessionReader); this commit only widens the write/storage
  type plumbing.
- SnapshotCatalog entry layout: size field 4 → 8 bytes (EntrySize 101 →
  105). Format break: existing on-disk catalogs are not readable by the
  new code. Acceptable on this branch; pre-release migration if needed
  later.
- MemoryArenaManager (test-only, byte[]-backed) accepts long but
  checked-casts to int internally — its arenas remain ≤2 GiB.

+++ src/Nethermind/Nethermind.State.Flat.Test/BSearchIndex/BSearchIndexTests.cs
@@ -129,7 +129,7 @@ public class BSearchIndexTests
             writer.AddKey(key, valBuf);
         }
         writer.FinalizeNode();
-        int written = bufWriter.Written;
+        int written = (int)bufWriter.Written;
 
         Assert.That(Convert.ToHexString(output[..written]), Is.EqualTo(expectedHex));
 
@@ -174,7 +174,7 @@ public class BSearchIndexTests
             writer.AddKey(Convert.FromHexString(sepHex), valBuf);
         }
         writer.FinalizeNode();
-        int written = bufWriter.Written;
+        int written = (int)bufWriter.Written;
 
         Assert.That(Convert.ToHexString(output[..written]), Is.EqualTo(expectedHex));
 
@@ -249,7 +249,7 @@ public class BSearchIndexTests
             writer.AddKey(key, valBuf);
         }
         writer.FinalizeNode();
-        int written = bufWriter.Written;
+        int written = (int)bufWriter.Written;
 
         Assert.That(Convert.ToHexString(output[..written]), Is.EqualTo(expectedHex));
 
@@ -333,7 +333,7 @@ public class BSearchIndexTests
             writer.AddKey(key, valBuf);
         }
         writer.FinalizeNode();
-        int written = bufWriter.Written;
+        int written = (int)bufWriter.Written;
 
         Assert.That(Convert.ToHexString(output[..written]), Is.EqualTo(expectedHex));
 
@@ -456,7 +456,7 @@ public class BSearchIndexTests
             writer.AddKey(sep.AsSpan(prefixLen), valBuf);
         }
         writer.FinalizeNode();
-        int written = w.Written;
+        int written = (int)w.Written;
 
         // Control node: same data without the prefix optimization (full-length keys,
         // no commonKeyPrefix passed). Demonstrates the size win.
@@ -558,7 +558,7 @@ public class BSearchIndexTests
         writer.AddKey(sepBuffer.AsSpan(2, 2), valBuf);
         writer.FinalizeNode();
 
-        BSearchIndexReader reader = BSearchIndexReader.ReadFromEnd(output, w.Written);
+        BSearchIndexReader reader = BSearchIndexReader.ReadFromEnd(output, (int)w.Written);
         Assert.That(reader.Metadata.HasCommonKeyPrefix, Is.False);
         Assert.That(reader.CommonKeyPrefix.Length, Is.EqualTo(0));
     }
@@ -600,7 +600,7 @@ public class BSearchIndexTests
         }
         writer.FinalizeNode();
 
-        BSearchIndexReader reader = BSearchIndexReader.ReadFromEnd(output, w.Written);
+        BSearchIndexReader reader = BSearchIndexReader.ReadFromEnd(output, (int)w.Written);
 
         // For each stored key plus a synthetic "between" probe, the two paths must agree.
         try
